requirements.txt
ccxt
python-dotenv
asyncio
requests